refactor: unify map pages onto a shared MapShell
