+++ src/pages/velocityHeatmap/index.tsx
@@ -1,11 +1,10 @@
-import { OpenInFullRounded } from '@mui/icons-material'
-import { FormControlLabel, IconButton, Radio, RadioGroup, Stack, Typography } from '@mui/material'
-import React, { useCallback, useContext, useRef, useState } from 'react'
+import { FormControlLabel, Radio, RadioGroup, Stack, Typography } from '@mui/material'
+import React, { useContext, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { AttributionControl, MapContainer, TileLayer, ZoomControl } from 'react-leaflet'
+import { TileLayer } from 'react-leaflet'
 import dayjs, { ISRAEL_TIMEZONE, toIsraelTimezone } from 'src/dayjs'
-import { useConstrainedFloatingButton } from 'src/hooks/useConstrainedFloatingButton'
 import { GlobalSearchContext } from 'src/model/globalState'
+import { MapShell } from 'src/pages/components/map-related/MapShell'
 import { DateNavigator } from '../components/dateNavigator/DateNavigator'
 import { DateSelector } from '../components/DateSelector'
 import { PageContainer } from '../components/PageContainer'
@@ -21,10 +20,7 @@ const VIS_MODES = [
 const DEFAULT_ZOOM_LEVEL = 10
 
 const VelocityHeatmapPage: React.FC = () => {
-  const [isExpanded, setIsExpanded] = useState<boolean>(false)
-  const toggleExpanded = useCallback(() => setIsExpanded((expanded) => !expanded), [])
-  const { t, i18n } = useTranslation()
-  const isRtl = i18n.dir() === 'rtl'
+  const { t } = useTranslation()
 
   const { search, setSearch } = useContext(GlobalSearchContext)
   const dateDayjs = dayjs.tz(search.date, ISRAEL_TIMEZONE)
@@ -33,17 +29,12 @@ const VelocityHeatmapPage: React.FC = () => {
   const [min, setMin] = useState(0)
   const [max, setMax] = useState(1)
 
-  const mapContainerRef = useRef<HTMLDivElement>(null)
-  const buttonRef = useRef<HTMLButtonElement>(null)
-
   const handleDateChange = (time: dayjs.Dayjs | null) => {
     setSearch((current) => ({
       ...current,
       date: toIsraelTimezone(time ?? dayjs()).format('YYYY-MM-DD'),
     }))
   }
-
-  useConstrainedFloatingButton(mapContainerRef, buttonRef, isExpanded)
 
   return (
     <PageContainer>
@@ -72,37 +63,24 @@ const VelocityHeatmapPage: React.FC = () => {
         ))}
       </RadioGroup>
 
-      <div ref={mapContainerRef} className={`map-info ${isExpanded ? 'expanded' : 'collapsed'}`}>
-        <IconButton
-          ref={buttonRef}
-          color="primary"
-          className="expand-button"
-          onClick={toggleExpanded}>
-          <OpenInFullRounded fontSize="large" />
-        </IconButton>
-        <MapContainer
-          center={[29.65, 34.6]}
-          zoom={DEFAULT_ZOOM_LEVEL}
-          scrollWheelZoom={true}
-          zoomControl={false}
-          attributionControl={false}
-          style={{ height: '100%', width: '100%' }}>
-          <ZoomControl position={isRtl ? 'topleft' : 'topright'} />
-          <AttributionControl position={isRtl ? 'bottomright' : 'bottomleft'} prefix={false} />
-          <TileLayer
-            attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-            url="https://tile-a.openstreetmap.fr/hot/{z}/{x}/{y}.png"
-          />
-          <VelocityHeatmapRectangles
-            visMode={visMode}
-            setMinMax={(min, max) => {
-              setMin(min)
-              setMax(max)
-            }}
-          />
-          <VelocityHeatmapLegend visMode={visMode} min={min} max={max} />
-        </MapContainer>
-      </div>
+      <MapShell
+        center={[29.65, 34.6]}
+        zoom={DEFAULT_ZOOM_LEVEL}
+        scrollWheelZoom={true}
+        style={{ height: '100%', width: '100%' }}>
+        <TileLayer
+          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+          url="https://tile-a.openstreetmap.fr/hot/{z}/{x}/{y}.png"
+        />
+        <VelocityHeatmapRectangles
+          visMode={visMode}
+          setMinMax={(min, max) => {
+            setMin(min)
+            setMax(max)
+          }}
+        />
+        <VelocityHeatmapLegend visMode={visMode} min={min} max={max} />
+      </MapShell>
     </PageContainer>
   )
 }
